+++ nym-vpn-apple/ServicesMacOS/Sources/GRPCManager/GRPCManager+SystemMessages.swift
@@ -1,4 +1,4 @@
-import NymVPNRpc
+import NymVPNLib
 
 extension GRPCManager {
     public func fetchCompatibleVersions() async throws -> (macOS: String?, core: String?) {