+++ nym-vpn-apple/ServicesMutual/Sources/ConnectionTypes/AccountSummary/Subscription.swift
@@ -1,14 +1,6 @@
-#if os(iOS)
 import NymVPNLib
-#elseif os(macOS)
-import NymVPNRpc
-#endif
 
-#if os(iOS)
 public typealias OuterSubscription = NymVPNLib.Subscription
-#elseif os(macOS)
-public typealias OuterSubscription = NymVPNRpc.Subscription
-#endif
 
 public struct Subscription: Codable {
     public let status: VpnSubscriptionStatus